# Performance of the file copy