@MrSerth 2024-09-26
This PR only contained time stamps to ease debugging with the file copy performance. Since we've gained additional monitoring capabilities in the mean time (including detailed Sentry logs) and haven't noticed any further issue with the performance of the file copy operation, we can close this PR.